Reorganize API into routers/services/auth layers
Split the monolithic app.py (~2100 lines) into focused modules:
- src/api/routers/: one file per resource (chat, threads, users, custom_areas, geometry, metadata)
- src/api/services/: business logic separated from HTTP concerns (auth, quota, chat)
- src/api/auth/dependencies.py: FastAPI auth dependencies (fetch_user_from_rw_api, require_auth, optional_auth)
- src/api/app.py: reduced to ~80 lines (app creation, middleware, router includes)

Also moves ThreadUpdateRequest into schemas.py and adds tests for the
previously untested geometry and metadata endpoints.
