+++ src/api/auth/__init__.py
@@ -1,3 +1,10 @@
+from .dependencies import fetch_user_from_rw_api, optional_auth, require_auth
 from .machine_user import MACHINE_USER_PREFIX, validate_machine_user_token
 
-__all__ = ["MACHINE_USER_PREFIX", "validate_machine_user_token"]
+__all__ = [
+    "MACHINE_USER_PREFIX",
+    "validate_machine_user_token",
+    "fetch_user_from_rw_api",
+    "require_auth",
+    "optional_auth",
+]
